aws-serverless/requirements.txt
# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Lambda Powertools
aws-lambda-powertools[all]>=2.32.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# JSON handling
orjson>=3.9.0

# HTTP client (for external APIs)
httpx>=0.26.0

# JWT handling (for custom authorizer)
python-jose[cryptography]>=3.3.0
cryptography>=41.0.0

# Caching
cachetools>=5.3.0

# Database (optional - for RDS integration)
# psycopg2-binary>=2.9.0  # PostgreSQL
# pymysql>=1.1.0          # MySQL

# Redis (optional - for ElastiCache)
# redis>=5.0.0
